Make build work with fsanitize=cilk
Lessons: -fsanitize needs to be applied to both CFLAGS and LDFLAGS. In addition,
there appears to not only be debug/release, but also lib/console configs. The
default build type is debug + console. So, I needed to add -fsanitize to the
console LDFLAGS. I also included the cilktools include folder but I don't know if
this is necessary.